Created GraphBuilder and NodeUtilities
GraphBuilder uses the ValidLocationCalculator to create a bunch of nodes
corresponding to the reachable locations in the level. Each Node was
modified to include a list of its neighbors which the GraphBuilder
populates appropriately and returns the root (StartLocation) node.

The NodeUtilities class adds extension methods which can be used on any
Node object to traverse the graph using either DFS or BFS.