Fixed bug in random level generation
The FurthestPointFinder has been fixed so that it won't accidently
select a start/end location pair that results in an unsolveable level.
This was happening because if the start tile that was chosen was
originally a wall and the path to the solution depended on that wall
being there, the level would potentially be unsolveable since the start
location works as a floor tile. This commit prevents that from happening
by setting the start location to a floor tile before generating the
level graph.